+++ zerver/lib/push_notifications.py
@@ -15,7 +15,7 @@ import orjson
 from aioapns.common import NotificationResult, PushType
 from django.conf import settings
 from django.db import transaction
-from django.db.models import F, Q
+from django.db.models import F, Q, QuerySet
 from django.db.models.functions import Lower
 from django.utils.timezone import now as timezone_now
 from django.utils.translation import gettext as _
@@ -42,6 +42,9 @@ from zerver.lib.exceptions import ErrorCode, JsonableError, MissingRemoteRealmEr
 from zerver.lib.message import access_message_and_usermessage, direct_message_group_users
 from zerver.lib.notification_data import get_mentioned_user_group
 from zerver.lib.remote_server import (
+    PushNotificationBouncerError,
+    PushNotificationBouncerRetryLaterError,
+    PushNotificationBouncerServerError,
     record_push_notifications_recently_working,
     send_json_to_push_bouncer,
     send_server_data_to_push_bouncer,
@@ -1463,9 +1466,14 @@ def get_encrypted_data(payload_data_to_encrypt: dict[str, Any], public_key_str: 
 def send_push_notifications(
     user_profile: UserProfile,
     payload_data_to_encrypt: dict[str, Any],
+    test_notification_to_push_devices: QuerySet[PushDevice] | None = None,
 ) -> None:
-    # Uses 'zerver_pushdevice_user_bouncer_device_id_idx' index.
-    push_devices = PushDevice.objects.filter(user=user_profile, bouncer_device_id__isnull=False)
+    if test_notification_to_push_devices is not None:
+        assert len(test_notification_to_push_devices) != 0
+        push_devices = test_notification_to_push_devices
+    else:
+        # Uses 'zerver_pushdevice_user_bouncer_device_id_idx' index.
+        push_devices = PushDevice.objects.filter(user=user_profile, bouncer_device_id__isnull=False)
 
     if len(push_devices) == 0:
         logger.info(
@@ -1475,6 +1483,7 @@ def send_push_notifications(
         return
 
     is_removal = payload_data_to_encrypt["type"] == "remove"
+    is_test_notification = payload_data_to_encrypt["type"] == "test"
 
     # Note: The "Final" qualifier serves as a shorthand
     # for declaring that a variable is effectively Literal.
@@ -1549,6 +1558,12 @@ def send_push_notifications(
             acting_user=None,
         )
         do_set_push_notifications_enabled_end_timestamp(user_profile.realm, None, acting_user=None)
+
+        if is_test_notification:
+            # Propagate the exception to the caller to notify the client
+            # about the error while attempting to send test push notification.
+            raise e
+
         return
 
     # Handle success response data
@@ -1596,6 +1611,12 @@ def send_push_notifications(
         )
         if can_push:
             record_push_notifications_recently_working()
+
+    if is_test_notification and len(push_requests) == len(delete_device_ids):
+        # While sending test push notification, the bouncer reported
+        # that there's no active registered push device. Inform the
+        # same to the client.
+        raise NoActivePushDeviceError
 
 
 def handle_push_notification(user_profile_id: int, missed_message: dict[str, Any]) -> None:
@@ -1797,6 +1818,39 @@ def send_test_push_notification(user_profile: UserProfile, devices: list[PushDev
     )
 
 
+def send_e2ee_test_push_notification(
+    user_profile: UserProfile, push_devices: QuerySet[PushDevice]
+) -> None:
+    payload_data_to_encrypt = get_base_payload(user_profile, for_legacy_clients=False)
+    payload_data_to_encrypt["type"] = "test"
+    payload_data_to_encrypt["time"] = datetime_to_timestamp(timezone_now())
+
+    logger.info("Sending E2EE test push notification for user %s", user_profile.id)
+
+    try:
+        send_push_notifications(
+            user_profile, payload_data_to_encrypt, test_notification_to_push_devices=push_devices
+        )
+    except PushNotificationBouncerServerError:
+        # 5xx error response from bouncer server
+        raise InternalBouncerServerError
+    except PushNotificationBouncerRetryLaterError:
+        # Network error
+        raise FailedToConnectBouncerError
+    except (
+        # Need to resubmit realm info - `manage.py register_server`
+        MissingRemoteRealmError,
+        # Invalid credentials or unexpected status code
+        PushNotificationBouncerError,
+        # Plan doesn't allow sending push notifications
+        PushNotificationsDisallowedByBouncerError,
+    ):
+        # Server admins need to fix these set of errors, report them.
+        error_msg = f"Sending E2EE test push notification for user_id={user_profile.id} failed."
+        logger.error(error_msg)
+        raise PushNotificationAdminActionRequiredError
+
+
 class InvalidPushDeviceTokenError(JsonableError):
     code = ErrorCode.INVALID_PUSH_DEVICE_TOKEN
 
@@ -1858,3 +1912,56 @@ def get_push_devices(user_profile: UserProfile) -> dict[str, PushDeviceInfoDict]
         str(row.push_account_id): {"status": row.status, "error_code": row.error_code}
         for row in rows
     }
+
+
+class NoActivePushDeviceError(JsonableError):
+    code = ErrorCode.NO_ACTIVE_PUSH_DEVICE
+
+    def __init__(self) -> None:
+        pass
+
+    @staticmethod
+    @override
+    def msg_format() -> str:
+        return _("No active registered push device")
+
+
+class FailedToConnectBouncerError(JsonableError):
+    http_status_code = 502
+    code = ErrorCode.FAILED_TO_CONNECT_BOUNCER
+
+    def __init__(self) -> None:
+        pass
+
+    @staticmethod
+    @override
+    def msg_format() -> str:
+        return _("Network error while connecting to Zulip push notification service.")
+
+
+class InternalBouncerServerError(JsonableError):
+    http_status_code = 502
+    code = ErrorCode.INTERNAL_SERVER_ERROR_ON_BOUNCER
+
+    def __init__(self) -> None:
+        pass
+
+    @staticmethod
+    @override
+    def msg_format() -> str:
+        return _("Internal server error on Zulip push notification service, retry later.")
+
+
+class PushNotificationAdminActionRequiredError(JsonableError):
+    http_status_code = 403
+    code = ErrorCode.ADMIN_ACTION_REQUIRED
+
+    def __init__(self) -> None:
+        pass
+
+    @staticmethod
+    @override
+    def msg_format() -> str:
+        return _(
+            "Push notification configuration issue on server, contact the server administrator or retry later."
+        )